chore(user-interviews): move contracts above facade to fix layering
Reviewer pointed out the inversion: logic.py was importing from
facade/contracts.py, which makes the inner layer depend on the outer
one. No cycle today (contracts has no imports from logic) but fragile.

Move contracts.py one level up to backend/contracts.py so logic.py and
facade/api.py both import it from a neutral sibling. facade/__init__.py
keeps re-exporting IntervieweeIdentity so external callers continue to
import it via the facade package.

Generated-By: PostHog Code
Task-Id: 639c774a-3c16-4b58-94a4-35682c62ade2
