+++ products/user_interviews/backend/facade/__init__.py
@@ -4,8 +4,8 @@ Facade for user_interviews product.
 This module provides the public interface for other products to interact with user interviews.
 """
 
+from ..contracts import IntervieweeIdentity
 from .api import has_replied, parse_interviewee_identifier
-from .contracts import IntervieweeIdentity
 
 __all__ = [
     "IntervieweeIdentity",
